Cleanup fullnode sync error handling
