--- crates/common/src/error.rs
@@ -1,15 +1,0 @@
-#[derive(Debug)]
-pub enum SyncError {
-    // Should not retry in this case
-    SequencerCommitmentNotFound([u8; 32]),
-    SequencerCommitmentWithIndexNotFound(u32),
-    SequencerCommitmentMissingForProof(u32),
-    UnknownL1Hash,
-    Error(anyhow::Error),
-}
-
-impl From<anyhow::Error> for SyncError {
-    fn from(e: anyhow::Error) -> Self {
-        Self::Error(e)
-    }
-}
