+++ crates/common/src/lib.rs
@@ -6,7 +6,6 @@ pub mod cache;
 pub mod config;
 pub mod da;
 pub mod db_migrations;
-pub mod error;
 pub mod l2;
 pub mod rpc;
 pub mod utils;
