+++ crates/fullnode/src/lib.rs
@@ -22,6 +22,7 @@ use tokio::sync::{broadcast, Mutex};
 
 pub mod da_block_handler;
 pub mod db_migrations;
+mod error;
 mod l2_syncer;
 mod metrics;
 pub mod rpc;
